Enhancements and Bugs fix: - bug fix: ELL throws exception when two PPT presentations are opened with ELL - bug fix: ELL throws exception when PPT closed with unsaved changes with ELL open - bug fix: incorrect callout / caption position during syncing. - bug fix: caption not removed when self-explanation item is deleted - enhancement: voice not generated/synced if caption text is empty - enhancement: change `slideselectionchange` behavior of ELL so that task pane do not close automatically - enhancement: check caption text/ callout text and uncheck isVoice/isCaption/isCallout if text is empty

+++ PowerPointLabs/PowerPointLabs/ELearningLab/Service/StorageService/AudioSettingStorageService.cs
@@ -33,7 +33,10 @@ namespace PowerPointLabs.ELearningLab.Service
             {
                 Directory.CreateDirectory(GetELearningLabStoragePath());
             }
-            File.Delete(GetAudioSettingFilePath());
+            if (File.Exists(GetAudioSettingFilePath()))
+            {
+                File.Delete(GetAudioSettingFilePath());
+            }
             Dictionary<string, string> audioSetting = new Dictionary<string, string>()
             {
                 { "voiceName", AudioSettingService.selectedVoice.ToString() }
